Remove `include/grpcpp/impl/codegen/call_op_set_interface.h`
