+++ CMakeLists.txt
@@ -3299,6 +3299,7 @@ foreach(_hdr
   include/grpcpp/health_check_service_interface.h
   include/grpcpp/impl/call.h
   include/grpcpp/impl/call_hook.h
+  include/grpcpp/impl/call_op_set_interface.h
   include/grpcpp/impl/channel_argument_option.h
   include/grpcpp/impl/client_unary_call.h
   include/grpcpp/impl/codegen/async_generic_service.h
@@ -3970,6 +3971,7 @@ foreach(_hdr
   include/grpcpp/health_check_service_interface.h
   include/grpcpp/impl/call.h
   include/grpcpp/impl/call_hook.h
+  include/grpcpp/impl/call_op_set_interface.h
   include/grpcpp/impl/channel_argument_option.h
   include/grpcpp/impl/client_unary_call.h
   include/grpcpp/impl/codegen/async_generic_service.h
