+++ src/cpp/server/server_context.cc
@@ -42,8 +42,8 @@
 #include <grpcpp/completion_queue.h>
 #include <grpcpp/ext/call_metric_recorder.h>
 #include <grpcpp/impl/call.h>
+#include <grpcpp/impl/call_op_set_interface.h>
 #include <grpcpp/impl/codegen/call_op_set.h>
-#include <grpcpp/impl/codegen/call_op_set_interface.h>
 #include <grpcpp/impl/codegen/callback_common.h>
 #include <grpcpp/impl/codegen/completion_queue_tag.h>
 #include <grpcpp/impl/codegen/interceptor_common.h>
